refactor (webapi): Move more JSON-LD logic into lower-level case classes.
