+++ webapi/src/main/scala/org/knora/webapi/messages/v2/responder/KnoraResponseV2.scala
@@ -637,133 +637,7 @@ case class ReadNamedGraphsV2(namedGraphs: Set[IRI]) extends KnoraResponseV2 {
     def toXML = ???
 }
 
-/**
-  * Return information about ontology entities.
-  *
-  * @param ontologies      named graphs and their resource classes.
-  * @param resourceClasses information about resource classes.
-  * @param properties      information about properties.
-  * @param userLang        the preferred language in which the information should be returned.
-  *
-  */
-case class ReadEntityDefinitionsV2(ontologies: Map[IRI, Set[IRI]] = Map.empty[IRI, Set[IRI]],
-                                   resourceClasses: Map[IRI, ResourceEntityInfoV2] = Map.empty[IRI, ResourceEntityInfoV2],
-                                   properties: Map[IRI, PropertyEntityInfoV2] = Map.empty[IRI, PropertyEntityInfoV2], userLang: String) extends KnoraResponseV2 {
 
-
-    def toJsonLDWithValueObject(settings: SettingsImpl): String = {
-
-        val context = new util.HashMap[String, String]()
-        context.put(OntologyConstants.KnoraApi.KnoraApiOntologyLabel, OntologyConstants.KnoraApiV2WithValueObject.KnoraApiV2PrefixExpansion)
-        context.put("rdfs", "http://www.w3.org/2000/01/rdf-schema#")
-        context.put("rdf", "http://www.w3.org/1999/02/22-rdf-syntax-ns#")
-        context.put("owl", "http://www.w3.org/2002/07/owl#")
-
-        val json: util.HashMap[String, Object] = new util.HashMap[String, Object]()
-
-        // ontologies with their resource classes
-
-        val ontos: util.Map[IRI, util.List[IRI]] = ontologies.map {
-            case (namedGraphIri: IRI, resourceClassIris: Set[IRI]) =>
-
-                val resClassIris = seqAsJavaList(resourceClassIris.toSeq.map {
-                    resClass =>
-                        InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(resClass, () => throw InconsistentTriplestoreDataException(s"internal resource class Iri $resClass could not be converted to knora-api v2 with value object entity Iri"))
-                })
-
-                InputValidation.internalOntologyIriToApiV2WithValueObjectOntologyIri(namedGraphIri, () => throw InconsistentTriplestoreDataException(s"internal ontology Iri could not be converted to knora-api v2 with value object ontology Iri")) -> resClassIris
-
-        }.asJava
-
-        json.put(OntologyConstants.KnoraApiV2WithValueObject.HasOntologiesWithResourceClasses, ontos)
-
-        // resource classes
-
-        val resClasses: util.Map[IRI, util.Map[IRI, Object]] = resourceClasses.map {
-            case (resClassIri: IRI, resourceEntity: ResourceEntityInfoV2) =>
-
-                val cardinalities = resourceEntity.cardinalities.map {
-
-                    case (propertyIri: IRI, cardinality: Cardinality.Value) =>
-
-                        val prop2card: (IRI, Int) = cardinality match {
-                            case Cardinality.MayHaveMany =>
-                                OntologyConstants.Owl.MinCardinality -> 0
-
-                            case Cardinality.MayHaveOne =>
-                                OntologyConstants.Owl.MaxCardinality -> 1
-
-                            case Cardinality.MustHaveOne =>
-                                OntologyConstants.Owl.Cardinality -> 1
-
-                            case Cardinality.MustHaveSome =>
-                                OntologyConstants.Owl.MinCardinality -> 1
-                        }
-
-                        Map(
-                            "@type" -> OntologyConstants.Owl.Restriction,
-                            OntologyConstants.Owl.OnProperty -> InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(propertyIri, () => throw InconsistentTriplestoreDataException(s"internal property Iri $propertyIri could not be converted to knora-api v2 with value object property Iri")),
-                            prop2card
-                        ).asJava
-                }.toSeq.asJava
-
-                val resourceIcon: Option[(IRI, String)] = resourceEntity.getPredicateAndObject(OntologyConstants.KnoraBase.ResourceIcon, settings) match {
-                    case Some((hasResIcon, resIcon)) =>
-                        Some(InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(hasResIcon, () => throw InconsistentTriplestoreDataException(s"internal property $hasResIcon could not be converted to knora-api v2 with value object property Iri")) -> resIcon)
-                    case None => None
-                }
-
-                InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(
-                    resClassIri,
-                    () => throw InconsistentTriplestoreDataException(s"internal resource class Iri $resClassIri could not be converted to a knora-api v2 with value object resource class Iri")
-                ) -> (Map(
-                    "@id" -> InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(resourceEntity.resourceClassIri, () => throw InconsistentTriplestoreDataException("")),
-                    OntologyConstants.KnoraApiV2WithValueObject.BelongsToOntology -> InputValidation.internalOntologyIriToApiV2WithValueObjectOntologyIri(resourceEntity.ontologyIri, () => throw InconsistentTriplestoreDataException(s"internal ontology Iri ${resourceEntity.ontologyIri} could not be converted to knora-api v2 with value object ontology Iri")),
-                    "@type" -> OntologyConstants.Owl.Class, // TODO: does this need to be coming from the triplestore?
-                    OntologyConstants.Rdfs.SubClassOf -> cardinalities
-                ) ++ resourceIcon
-                    ++ resourceEntity.getPredicateAndObject(OntologyConstants.Rdfs.Label, settings, Some(userLang))
-                    ++ resourceEntity.getPredicateAndObject(OntologyConstants.Rdfs.Comment, settings, Some(userLang))).asJava
-        }.asJava
-
-        json.put(OntologyConstants.KnoraApiV2WithValueObject.HasResourceClasses, resClasses)
-
-        // properties
-
-        val props: util.Map[IRI, util.Map[IRI, String]] = properties.map {
-            case (propIri: IRI, propEntity: PropertyEntityInfoV2) =>
-
-                val objectClassConstraint: Option[(IRI, String)] = propEntity.getPredicateAndObject(OntologyConstants.KnoraBase.ObjectClassConstraint, settings) match {
-                    case Some((objectClassConstrPred: IRI, objectClassConstrObj: IRI)) =>
-                        Some(InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(
-                            objectClassConstrPred,
-                            () => throw InconsistentTriplestoreDataException(s"internal property $objectClassConstrPred could not be converted to knora-api v2 with value object property Iri")) -> InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(objectClassConstrObj, () => throw InconsistentTriplestoreDataException("")
-                        ))
-
-                    case None => None
-                }
-
-                InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(propIri, () => throw InconsistentTriplestoreDataException(s"internal property $propIri could not be converted to knora-api v2 with value object property Iri")) -> (Map(
-                    "@id" -> InputValidation.internalEntityIriToApiV2WithValueObjectEntityIri(propEntity.propertyIri, () => throw InconsistentTriplestoreDataException(s"internal property $propIri could not be converted to knora-api v2 with value object property Iri")),
-                    OntologyConstants.KnoraApiV2WithValueObject.BelongsToOntology -> InputValidation.internalOntologyIriToApiV2WithValueObjectOntologyIri(propEntity.ontologyIri, () => throw InconsistentTriplestoreDataException(s"internal ontology ${propEntity.ontologyIri} could not be converted to knora-api v2 with value object ontology Iri"))
-                ) ++ propEntity.getPredicateAndObject(OntologyConstants.Rdf.Type, settings)
-                    ++ propEntity.getPredicateAndObject(OntologyConstants.Rdfs.Label, settings, Some(userLang))
-                    ++ propEntity.getPredicateAndObject(OntologyConstants.Rdfs.Comment, settings, Some(userLang))
-                    ++ objectClassConstraint).asJava
-        }.asJava
-
-        json.put(OntologyConstants.KnoraApiV2WithValueObject.HasProperties, props)
-
-        val compacted = JsonLdProcessor.compact(json, context, new JsonLdOptions())
-
-        JsonUtils.toPrettyString(compacted)
-
-
-    }
-
-    def toXML = ???
-
-}
 
 object ReadResourceUtil {
     /**
